Split the solver pair files -> solver, state, and time integrator pair files for clarity and better paritioning
