+++ src/rigid_pendulum_poc/CMakeLists.txt
@@ -4,5 +4,7 @@ target_sources(${oturb_lib_name}
     # diagnostics.cpp
     # io.cpp
     solver.cpp
+    state.cpp
+    time_integrator.cpp
     # IOManager.cpp
 )
