+++ tests/unit_tests/rigid_pendulum_poc/test_generalized_alpha_solver.cpp
@@ -3,6 +3,8 @@
 #include <gtest/gtest.h>
 
 #include "src/rigid_pendulum_poc/solver.h"
+#include "src/rigid_pendulum_poc/state.h"
+#include "src/rigid_pendulum_poc/time_integrator.h"
 #include "tests/unit_tests/rigid_pendulum_poc/test_utilities.h"
 
 namespace openturbine::rigid_pendulum::tests {
